Move remaining JS files to Typescript, second iteration
Moved DropDownButton, ProgressBar, Navbar, Heading, Gauge, and
DeviceStatusGauge to Typescript. Updated storybook to render
bordered items when dropdown list is expanded.

Also contains a bugfix for Navbar not rendering correctly on mobile.

Connects-to: #44
Change-type: patch
