+++ src/components/Gauge.tsx
@@ -219,4 +219,6 @@ class StatusGauge extends React.Component<GaugeProps, {}> {
 	}
 }
 
-export default compose(withTheme, asRendition)(StatusGauge);
+export default compose(withTheme, asRendition)(
+	StatusGauge,
+) as React.ComponentClass<GaugeProps>;
